requirements.txt
astor==0.8.1
autograd==1.6.2
autograd-gamma==0.5.0
cached-property==1.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
cycler==0.11.0
fonttools==4.38.0
formulaic==0.6.4
future==0.18.3
graphlib-backport==1.0.3
idna==3.4
interface-meta==1.3.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lifelines==0.27.7
MarkupSafe==2.1.3
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.5
olefile==0.46
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pip==22.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
rdkit==2022.3.3
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
threadpoolctl==3.1.0
torch-geometric==2.0.4
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
wheel==0.38.4
wrapt==1.15.0
